implementação com email funcionando após adição de dominio
