+++ backend/.env.example
@@ -7,8 +7,8 @@ JWT_SECRET=dev-only-change-me-with-at-least-32-chars
 GOOGLE_CLIENT_ID=replace-with-google-client-id
 GOOGLE_CLIENT_SECRET=replace-with-google-client-secret
 FRONTEND_URL=http://localhost:4200
-BREVO_API_KEY=replace-with-brevo-api-key
-EMAIL_FROM=seu-remetente-verificado@gmail.com
+RESEND_API_KEY=replace-with-resend-api-key
+EMAIL_FROM=no-reply@seu-dominio.com
 EMAIL_FROM_NAME=BudgetFlow
 
 # Production
@@ -23,6 +23,6 @@ EMAIL_FROM_NAME=BudgetFlow
 # GOOGLE_CLIENT_ID=replace-with-google-client-id
 # GOOGLE_CLIENT_SECRET=replace-with-google-client-secret
 # FRONTEND_URL=https://app.budgetflow.com
-# BREVO_API_KEY=replace-with-brevo-api-key
-# EMAIL_FROM=seu-remetente-verificado@gmail.com
+# RESEND_API_KEY=replace-with-resend-api-key
+# EMAIL_FROM=no-reply@seu-dominio.com
 # EMAIL_FROM_NAME=BudgetFlow
